Move secrets to config dir
