+++ src/util/jest/config.ts
@@ -4,7 +4,7 @@ import { agenda } from '@/util/agenda';
 import { mockClear } from './mock';
 import { logger } from '@/util/logger';
 import { getProvider, NetworkProvider } from '@/util/network';
-import { POOL_REGISTRY_ADDRESS } from '@/util/secrets';
+import { POOL_REGISTRY_ADDRESS } from '@/config/secrets';
 import { poll } from './polling';
 
 export async function beforeAllCallback() {
